+++ packages/build-config/CHANGELOG.md
@@ -1,9 +1,14 @@
+## [Unreleased]
+
+### Features
+
+* **build-config:** add tsdown configuration utilities
+
 # 0.1.0 (2025-07-06)
 
 
 ### Features
 
 * **build-config:** add tsup configuration utilities ([b9ce9df](https://github.com/chengpeiquan/bassist/commit/b9ce9df24cec3ba55f52f7469a07a7d29e229c66))
-